+++ hp/hp.cabal
@@ -26,6 +26,7 @@ source-repository head
 library
   exposed-modules:
       AbstractSyntaxTree
+      ErrorHandler
       HpParser
       Lexer
       SemanticAnalyzer
